+++ templates/index.php
@@ -5,9 +5,24 @@
             <?=excerpt($post['content'])?>
         </article>
     <?php endforeach ?>
-    <nav>
-        <?php if(isset($previous_url)):?><a style="float: left;" href="<?=$previous_url?>">&laquo; <?=$messages['previous_page']?></a><?php endif?>
-        <?php if(isset($next_url)):?><a style="float: right;" href="<?=$next_url?>"><?=$messages['next_page']?> &raquo;</a><?php endif?>
+    <nav id="pagination" aria-label="<?=$messages['aria-pagination']?> }}">
+        <ul>
+        <?php if(isset($previous_url)): ?>
+            <li><a href="<?= $previous_url?>">&laquo; <?= $messages['previous_page']?></a></li>
+        <?php endif?>
+        <?php foreach($pagination as $p): ?>
+            <?php if(null === $p):?>
+                <li>. . .</li>
+            <?php elseif($p === $current_page):?>
+                <li class="current"><?=$p?></li>
+            <?php else:?>
+                <li><a href="<?= index_url($p) ?>"><?=$p?></a></li>
+            <?php endif?>
+        <?php endforeach?>
+        <?php if(isset($next_url)):?>
+            <li><a href="<?=$next_url?>"><?=$messages['next_page']?> &raquo;</a></li>
+        <?php endif?>
+        </ul>
     </nav>
 <?php else: ?>
 <p>This motherfucking static blog is fucking empty.</p>